added gestures

+++ ionic/conFusion/www/js/controllers.js
@@ -233,7 +233,7 @@ angular.module('conFusion.controllers', [])
                     }])
 
 
-.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
 
     $scope.baseURL = baseURL;
     $scope.shouldShowDelete = false;
@@ -254,12 +254,27 @@ angular.module('conFusion.controllers', [])
         console.log($scope.shouldShowDelete);
     }
 
-    $scope.deleteFavorite = function (index) {
-        
-        favoriteFactory.deleteFromFavorites(index);
+     $scope.deleteFavorite = function (index) {
+
+        var confirmPopup = $ionicPopup.confirm({
+            title: 'Confirm Delete',
+            template: 'Are you sure you want to delete this item?'
+        });
+
+        confirmPopup.then(function (res) {
+            if (res) {
+                console.log('Ok to delete');
+                favoriteFactory.deleteFromFavorites(index);
+            } else {
+                console.log('Canceled delete');
+            }
+        });
+
         $scope.shouldShowDelete = false;
 
-    }}])
+    }
+
+}])
 
 .filter('favoriteFilter', function () {
     return function (dishes, favorites) {
